migrated the grpc server main func to a dedicated main package. testing manually
