+++ internal/gcpmanager/main.go
@@ -3,27 +3,24 @@ package gcpmanager
 
 import (
 	"context"
-	"log"
-	"net"
 	"time"
 
 	sci "github.com/substratusai/substratus/internal/sci"
 
 	storage "cloud.google.com/go/storage"
-	"google.golang.org/grpc"
 )
 
 // server implements the sci.ControllerServer interface.
-type server struct {
+type Server struct {
 	sci.UnimplementedControllerServer
-	storageClient *storage.Client
+	StorageClient *storage.Client
 }
 
 // CreateSignedURL generates a signed URL for a specified GCS bucket and object path.
-func (s *server) CreateSignedURL(ctx context.Context, req *sci.CreateSignedURLRequest) (*sci.CreateSignedURLResponse, error) {
+func (s *Server) CreateSignedURL(ctx context.Context, req *sci.CreateSignedURLRequest) (*sci.CreateSignedURLResponse, error) {
 	bucketName, objectName := req.GetBucketName(), req.GetObjectName()
 
-	bucket := s.storageClient.Bucket(bucketName)
+	bucket := s.StorageClient.Bucket(bucketName)
 	obj := bucket.Object(objectName)
 	if _, err := obj.Attrs(ctx); err != nil {
 		if err != storage.ErrObjectNotExist {
@@ -42,32 +39,13 @@ func (s *server) CreateSignedURL(ctx context.Context, req *sci.CreateSignedURLRe
 	}
 
 	// Create a signed URL
+	// TODO(bjb): Need to test this on a live pod with WI with the SA having roles/iam.serviceAccountTokenCreator.
+	// I've only used this API with a static credential previously: https://cloud.google.com/storage/docs/access-control/signing-urls-with-helpers#upload-object
+	// a few techniques here: https://stackoverflow.com/questions/62439257/how-to-generate-signed-urls-for-google-cloud-storage-objects-in-gke-go
 	url, err := storage.SignedURL(bucketName, objectName, opts)
 	if err != nil {
 		return nil, err
 	}
 
 	return &sci.CreateSignedURLResponse{Url: url}, nil
-}
-
-func main() {
-	// Create a storage client
-	storageClient, err := storage.NewClient(context.Background())
-	if err != nil {
-		log.Fatalf("failed to create storage client: %v", err)
-	}
-
-	s := grpc.NewServer()
-	sci.RegisterControllerServer(s, &server{
-		storageClient: storageClient,
-	})
-
-	lis, err := net.Listen("tcp", ":9090")
-	if err != nil {
-		log.Fatalf("failed to listen: %v", err)
-	}
-
-	if err := s.Serve(lis); err != nil {
-		log.Fatalf("failed to serve: %v", err)
-	}
 }
